- feat: 마이페이지 프로필 영역 ProfileComp.tsx로 분리 및 NicknameComp.tsx -> NicknameInputComp.tsx로 파일명 변경.
