+++ app/(pages)/mypage/page.tsx
@@ -1,8 +1,7 @@
 'use client';
 
-import Image from 'next/image';
 import styled from 'styled-components';
-import NicknameComp from '@/app/components/mypage/NicknameComp';
+import ProfileComp from '@/app/components/mypage/ProfileComp';
 import Tabs from '@/app/components/common/TabsComp';
 import ReactionContent from '@/app/components/mypage/ReactionContent';
 import ScrapContent from '@/app/components/mypage/ScrapContent';
@@ -16,42 +15,6 @@ const ContainerMain = styled.div`
   > *:not(:last-child) {
     margin-bottom: 24px;
   }
-`;
-
-const ProfileDiv = styled.div`
-  display: flex;
-  flex-directioin: row;
-`;
-
-const ProfileImageDiv = styled.div`
-  width: 140px;
-  height: 140px;
-  position: relative;
-  overflow: hidden;
-  margin-right: 24px;
-  border-radius: 50%;
-  background-color: #e0e0e0;
-`;
-
-const ProfileTextDiv = styled.div`
-  display: flex;
-  flex-direction: column;
-
-  > *:not(:last-child) {
-    margin-bottom: 12px;
-  }
-`;
-
-const TitleStrong = styled.strong`
-  font-size: 24px;
-`;
-
-const TagP = styled.p`
-  padding: 2px;
-  text-align: center;
-  font-size: 24px;
-  border-radius: 8px;
-  background-color: #e0e0e0;
 `;
 
 const tabList = [
@@ -72,18 +35,7 @@ const tabList = [
 export default function MyPage() {
   return (
     <ContainerMain>
-      <ProfileDiv>
-        <ProfileImageDiv>
-          <Image src={'/next.svg'} alt="프로필 이미지" fill={true} />
-        </ProfileImageDiv>
-
-        <ProfileTextDiv>
-          <TitleStrong>프로필</TitleStrong>
-          <NicknameComp />
-          <TagP>태그</TagP>
-        </ProfileTextDiv>
-      </ProfileDiv>
-
+      <ProfileComp />
       <Tabs tabList={tabList} />
     </ContainerMain>
   );
